+++ main-server/src/core/as/event_handlers.js
@@ -49,6 +49,8 @@ export async function handleMessageFromQueue(message, nodeId = config.nodeId) {
     messageJSON: message,
   });
 
+  common.incrementProcessingInboundMessagesCount();
+
   const requestId = message.request_id;
   try {
     if (message.type === privateMessageType.DATA_REQUEST) {
@@ -103,6 +105,8 @@ export async function handleMessageFromQueue(message, nodeId = config.nodeId) {
       error: err,
       requestId,
     });
+  } finally {
+    common.decrementProcessingInboundMessagesCount();
   }
 }