+++ main-server/src/http_server/routes/server_info.js
@@ -20,18 +20,23 @@
  *
  */
 
-import path from 'path';
-import fs from 'fs';
+import express from 'express';
 
 import * as tendermint from '../../tendermint';
 import * as tendermintNdid from '../../tendermint/ndid';
+import * as common from '../../core/common';
+import * as mq from '../../mq';
 import * as mqService from '../../mq/grpc_client';
+import * as callbackUtil from '../../utils/callback';
+import * as externalCryptoService from '../../utils/external_crypto_service';
 import * as config from '../../config';
 import logger from '../../logger';
 
-let version;
+import { version } from '../../version';
 
-export default async function getInfo(req, res, next) {
+const router = express.Router();
+
+router.get('/info', async function getInfo(req, res, next) {
   try {
     let nodeInfo;
     if (tendermint.connected) {
@@ -80,18 +85,57 @@ export default async function getInfo(req, res, next) {
   } catch (error) {
     next(error);
   }
-}
+});
 
-fs.readFile(
-  path.join(__dirname, '..', '..', '..', '..', 'VERSION'),
-  'utf8',
-  (err, data) => {
-    if (err) {
-      logger.error({
-        message: 'Unable to read VERSION file',
-        err,
-      });
-    }
-    version = data;
-  }
-);
+router.get('/num_expected_txs', (req, res) => {
+  res.status(200).json({
+    n_expected_txs: tendermint.getExpectedTxsCount(),
+  });
+});
+
+router.get('/expected_txs', (req, res) => {
+  res.status(200).json({
+    expected_txs: tendermint.getExpectedTxHashes(),
+  });
+});
+
+// Outbound MQ messages that have not received ACK from destination
+router.get('/num_pending_outbound_mq_messages', (req, res) => {
+  res.status(200).json({
+    n_pending_outbound_mq_messages: mq.getPendingOutboundMessagesCount(),
+  });
+});
+
+// Pending callback calls that have not received response
+router.get('/num_pending_client_callbacks', (req, res) => {
+  res.status(200).json({
+    n_pending_client_callbacks: callbackUtil.getPendingCallbacksCount(),
+  });
+});
+
+// Pending callback to external decrypt and sign that have not received response
+router.get('/num_pending_external_crypto_callbacks', (req, res) => {
+  res.status(200).json({
+    n_pending_external_crypto_callbacks: externalCryptoService.getPendingCallbacksCount(),
+  });
+});
+
+router.get('/num_processing_blocks', (req, res) => {
+  res.status(200).json({
+    n_processing_blocks: tendermint.getProcessingBlocksCount(),
+  });
+});
+
+router.get('/processing_blocks', (req, res) => {
+  res.status(200).json({
+    processing_blocks: tendermint.getProcessingBlocks(),
+  });
+});
+
+router.get('/num_processing_inbound_messages', (req, res) => {
+  res.status(200).json({
+    n_processing_inbound_messages: common.getProcessingInboundMessagesCount(),
+  });
+});
+
+export default router;